Stage 3b: exact decimal money and weight (DAT-1, DAT-3)
Three representations of money fed into one another: whole-rupee INT on sale
and sale_details, and floating-point DOUBLE on customer.balance_amount and
every customer_ledger column. Binary floating point cannot hold decimal
currency exactly, so running balances drifted in a way no report could
reconcile. V2 gives the whole sale -> ledger -> customer chain DECIMAL columns
and the entities BigDecimal fields, and all ledger arithmetic now goes through
MoneyRules rather than double addition.

The weight problem was worse and is now repaired. sale_details
.total_kilogram_sale was INT while sale.kilograms is decimal, so every trip
header truncated its own total. Measured on production data: 1,833 of 2,228
trips did not tally with their line items and 9,838.521 kg was missing in
aggregate. V2 widens the column and recovers each header from its lines.

Only weight is recomputed. total_amount, total_payment_received and
total_bird_sale also disagree with their lines on 114, 97 and 107 trips, but
those are whole numbers on both sides, so a mismatch is not truncation and
could be wrong on either side; overwriting them would destroy information.
They are reported for investigation instead.

Seven trips still do not tally after the migration. All seven are headers with
zero sale lines, totalling 611 kg - their lines appear to have been deleted.
Left untouched deliberately.

createSingleSale now recalculates the amount and rejects a mismatch, matching
the bulk path from stage 3a.

Fixes a pre-existing bug in saveSaleDetails: setTotalAmount was called twice
when merging into an existing trip header, so the amount was added twice. This
likely accounts for some of the 114 amount mismatches above.

Verified against a full copy of production data (33,020 sales, 2,228 trips, 425
customers). Flyway baselined the copy at v1, applied V2, and the application
started with Hibernate ddl-auto=validate against the new BigDecimal entities -
the same upgrade path production will follow. Money totals were byte-identical
before and after (178,089,009.00 in sales; 14,352,997.00 in customer balances)
and sale line weights unchanged at 1,569,130.521 kg. The copy was dropped
afterwards; the real database is untouched and still has no flyway history.

Not yet exercised at runtime: backdated ledger recalculation under BigDecimal.
The copy carries only one ledger row, so that path had nothing to recalculate.
It needs testing once the ledger is backfilled.

Purchase, trading and dc_detail money columns are untouched here - including
purchase.diesel/driver_expense/hamali and dc_detail.kilograms/rate, which are
varchar(255). Their data is entirely numeric, so they can be converted, but as
a separate migration.

+++ src/main/java/com/app/dto/CustomerLedgerDTO.java
@@ -1,5 +1,6 @@
 package com.app.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -20,9 +21,9 @@ public class CustomerLedgerDTO {
     private TransactionType transactionType;
     private String referenceType;
     private Long referenceId;
-    private Double debitAmount;
-    private Double creditAmount;
-    private Double runningBalance;
+    private BigDecimal debitAmount;
+    private BigDecimal creditAmount;
+    private BigDecimal runningBalance;
     private String description;
     private String paymentMode;
     private LocalDateTime createdAt;
